Correcciones: Costos Extras y Reporte Orden Producción
Se modifico que sea posible agregar costos extras en todas las fases de
la cotización excepto en estatus “Finalizado”.

Se modifico la impresión de los campos: ancho, alto, ancho extendido y
alto extendido para que fueran flotantes en lugar de enteros.

+++ src/main/java/com/artiffex/lithomat/sistemaweb/businesstier/service/implementacion/CalificacionServiceImpl.java
@@ -1049,10 +1049,10 @@ public class CalificacionServiceImpl implements CalificacionService {
 				otttd.setMaquina(tipoTrabajoDetalle.getMaquina().getNombre());
 				otttd.setTipoPlaca(tipoTrabajoDetalle.getTipoPlaca().getDescripcion());
 				otttd.setDescripcion(tipoTrabajoDetalle.getDescripcion());
-				otttd.setAncho((int)tipoTrabajoDetalle.getAncho());
-				otttd.setAlto((int)tipoTrabajoDetalle.getAlto());
-				otttd.setAnchoExtendido((int)tipoTrabajoDetalle.getAnchoExtendido());
-				otttd.setAltoExtendido((int)tipoTrabajoDetalle.getAltoExtendido());
+				otttd.setAncho(tipoTrabajoDetalle.getAncho());
+				otttd.setAlto(tipoTrabajoDetalle.getAlto());
+				otttd.setAnchoExtendido(tipoTrabajoDetalle.getAnchoExtendido());
+				otttd.setAltoExtendido(tipoTrabajoDetalle.getAltoExtendido());
 				otttd.setTipoPapel(tipoTrabajoDetalleService.obtienePapelDescripcionBasica(tipoTrabajoDetalle.getIdTipoTrabajoDetalle()));
 				otttd.setNumeroPaginas(tipoTrabajoDetalle.getNumeroPaginasPublicacion());
 				otttd.setTamanioPublicacion(tipoTrabajoDetalle.getTamanioPublicacion().getNombre());
